8182554: Code for os::random() assumes long is 32 bits
And make updating the _rand_seed thread safe.

+++ hotspot/src/share/vm/prims/whitebox.cpp
@@ -227,8 +227,8 @@ static jint wb_stress_virtual_space_resize(size_t reserved_space_size,
     return 3;
   }
 
-  long seed = os::random();
-  tty->print_cr("Random seed is %ld", seed);
+  int seed = os::random();
+  tty->print_cr("Random seed is %d", seed);
   os::init_random(seed);
 
   for (size_t i = 0; i < iterations; i++) {
